Move libcontainer helper to executor package
