+++ drivers/shared/executor/libcontainer_nsenter_linux.go
@@ -1,4 +1,4 @@
-package main
+package executor
 
 import (
 	"os"
